Convert ProjectCard to typescript.
I removed the deleting code since it did nothing and was causing problems. This will be fixed in #379 anyways
